backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
plaid-python>=11.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pdfplumber>=0.11.0
python-multipart>=0.0.9
python-dateutil>=2.9.0
pytest>=8.0.0
httpx>=0.27.0
